# openAPI_specification/OpenAPI.yaml
openapi: "3.0.3"
info:
  title: Dummy Roblox Shop
  version: "1.0"
  description: A Codecademy exercise building for fun very basic online e-store - Roblox Shop API
servers:
  - url: https://api.server.test/v1
...
security:
  - BasicAuth: []
paths:
  /users/login:
    summary: User login
    description: This path is used for logging registered users in.
    post:
      tags:
        - Users
      summary: Log user in
      description: Enables user logging in. Requires user credentials passed as object
      operationId: login_user
      security: []
      requestBody:
        description: Passing User credentials
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCredentials"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/UserCredentials"
      responses:
        "303":
          description: User successfully logged in and redirected to profile page
          headers:
            Location:
              $ref: "#/components/headers/Location"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/register:
    summary: Allows user registration
    description: This path is used for registration of new users.
    post:
      tags:
        - Users
      summary: Register new user
      description: Insert new user personal data into DB for registration
      security: []
      operationId: add_user_profile
      requestBody:
        description: User registration data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BaseUser"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/BaseUser"
      responses:
        "303":
          description: User registration successful, user redirected to login page.
          headers:
            Location:
              $ref: "#/components/headers/Location"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/profile:
    summary: Get the user details
    description: This path is used for retreiving User Data for authorized users.
    get:
      tags:
        - Users
      summary: Retreive user profile
      description: Retreives the User related information from DB if the user is signed in.
      security:
        - BasicAuth: []
      operationId: get_user_profile
      responses:
        "200":
          description: User data retreived successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/update:
    summary: Allows user profile update
    description: This path is used for updating user information.
    post:
      tags:
        - Users
      summary: Update user records
      description: Enables update of User Profile related data if user is singed in.
      operationId: update_user_profile
      requestBody:
        description: Updated User data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "303":
          description: User record update successful, user redirected to profile page.
          headers:
            Location:
              $ref: "#/components/headers/Location"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/logout:
    summary: User logout
    description: This path is used for logging users out
    get:
      tags:
        - Users
      summary: Log user out
      description: Log out and redirects user to the login page
      security: []
      operationId: logout_user
      responses:
        "303":
          description: User successfully logged out and redirected to login page
          headers:
            Location:
              $ref: "#/components/headers/Location"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  schemas:
    BaseUser:
      type: object
      properties:
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        password:
          type: string
        status:
          type: string
      required:
        - firstname
        - lastname
        - email
        - password
        - status
    User:
      allOf:
        - type: object
          properties:
            id:
              type: integer
        - $ref: "#/components/schemas/BaseUser"
    UserUpdate:
      allOf:
        - $ref: "#/components/schemas/BaseUser"
        - type: object
          properties:
            newPassword:
              type: string
              nullable: true
            newPasswordCheck:
              type: string
              nullable: true
    UserCredentials:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
...
  headers:
    Location:
      description: The redirect URI
      schema:
        type: string
        example: /user/profile
  responses:
    InternalServerError:
      description: Error at server end.
      content:
        application/json:
          schema:
            type: object
            properties:
              msg:
                type: string
                description: Error message
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic